backend/requirements.txt
# ResearchHelper Backend Requirements

# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6  # For file uploads

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0  # For migrations
# Add database driver (e.g., psycopg2 for PostgreSQL, or asyncpg)

# Pydantic for validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Vector Database (choose one or multiple)
faiss-cpu>=1.7.4  # FAISS for local vector storage
# pinecone-client>=2.2.0
# qdrant-client>=1.7.0
# weaviate-client>=3.25.0

# LLM and Embeddings
openai>=1.3.0
anthropic>=0.7.0
# groq>=0.3.0  # When available
sentence-transformers>=2.2.0  # For local embeddings

# PDF Processing
PyPDF2>=3.0.0
# pdfplumber>=0.10.0
# pymupdf>=1.23.0

# DOI Fetching
requests>=2.31.0
# crossref-commons>=0.0.7

# Auth
PyJWT>=2.8.0
google-auth>=2.24.0
pymongo>=4.6.0
certifi>=2023.0.0  # SSL certificates for MongoDB Atlas

# Text Processing
nltk>=3.8.0
spacy>=3.7.0

# Graph Processing (optional)
networkx>=3.2.0
numpy>=1.24.0

# Email
resend>=2.0.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0  # For async HTTP requests

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0  # For test client

# Development
black>=23.11.0
ruff>=0.1.0
mypy>=1.7.0

